refactor(routes): share query string parsing helper
